更新单元测试文件,并修复BUG
