+++ docs/sphinx_md/appbuilder.core.components.retriever.md
@@ -9,6 +9,13 @@
       * [`BaiduVDBRetriever.name`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.BaiduVDBRetriever.name)
       * [`BaiduVDBRetriever.run()`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.BaiduVDBRetriever.run)
       * [`BaiduVDBRetriever.tool_desc`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.BaiduVDBRetriever.tool_desc)
+    * [`BaiduVDBVectorStoreIndex`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.BaiduVDBVectorStoreIndex)
+      * [`BaiduVDBVectorStoreIndex.add_segments()`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.BaiduVDBVectorStoreIndex.add_segments)
+      * [`BaiduVDBVectorStoreIndex.as_retriever()`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.BaiduVDBVectorStoreIndex.as_retriever)
+      * [`BaiduVDBVectorStoreIndex.client`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.BaiduVDBVectorStoreIndex.client)
+      * [`BaiduVDBVectorStoreIndex.from_params()`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.BaiduVDBVectorStoreIndex.from_params)
+      * [`BaiduVDBVectorStoreIndex.vdb_uri_prefix`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.BaiduVDBVectorStoreIndex.vdb_uri_prefix)
+    * [`TableParams`](appbuilder.core.components.retriever.baidu_vdb.md#appbuilder.core.components.retriever.baidu_vdb.baiduvdb_retriever.TableParams)
 * [appbuilder.core.components.retriever.bes package](appbuilder.core.components.retriever.bes.md)
   * [Submodules](appbuilder.core.components.retriever.bes.md#submodules)
   * [appbuilder.core.components.retriever.bes.bes_retriever module](appbuilder.core.components.retriever.bes.md#module-appbuilder.core.components.retriever.bes.bes_retriever)
@@ -17,6 +24,17 @@
       * [`BESRetriever.name`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESRetriever.name)
       * [`BESRetriever.run()`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESRetriever.run)
       * [`BESRetriever.tool_desc`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESRetriever.tool_desc)
+    * [`BESVectorStoreIndex`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex)
+      * [`BESVectorStoreIndex.add_segments()`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.add_segments)
+      * [`BESVectorStoreIndex.as_retriever()`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.as_retriever)
+      * [`BESVectorStoreIndex.base_es_url`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.base_es_url)
+      * [`BESVectorStoreIndex.create_index_mappings()`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.create_index_mappings)
+      * [`BESVectorStoreIndex.delete_all_segments()`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.delete_all_segments)
+      * [`BESVectorStoreIndex.es`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.es)
+      * [`BESVectorStoreIndex.from_segments()`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.from_segments)
+      * [`BESVectorStoreIndex.generate_id()`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.generate_id)
+      * [`BESVectorStoreIndex.get_all_segments()`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.get_all_segments)
+      * [`BESVectorStoreIndex.helpers`](appbuilder.core.components.retriever.bes.md#appbuilder.core.components.retriever.bes.bes_retriever.BESVectorStoreIndex.helpers)
 * [appbuilder.core.components.retriever.reranker package](appbuilder.core.components.retriever.reranker.md)
   * [Submodules](appbuilder.core.components.retriever.reranker.md#submodules)
   * [appbuilder.core.components.retriever.reranker.rerank module](appbuilder.core.components.retriever.reranker.md#module-appbuilder.core.components.retriever.reranker.rerank)
